fix: move generateLease to /components/lease, bug fixing
